fix: Enemies now reset out of view of the game board
I adjsuted the enemy this.x values for the reset function, such that
they move fully off of the board before resetting,
and reset to begin off of the left side of the board when re-instancing.
Previously, they would reset as soon as their right edge hit the
right edge of the board. Movement looks more natural now